+++ CHANGELOG.md
@@ -20,12 +20,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Response` class `size` attribute
 - `Prompt` class `size` attribute
 - `PromptTemplate` class `size` attribute
+- `examples` directory
 ### Changed
 - Validation bug fixed in `update_messages` method in `Session` class
 - Validation bug fixed in `from_json` method in `PromptTemplate`, `Response`, `Prompt`, and `Session` classes
 - `AI_STUDIO` render format modified
 - `Session` class messages status bug fixed
 - Test system modified
+- `README.md` updated
 ## [0.6] - 2025-05-05
 ### Added
 - `Response` class `id` property